pyproject.toml
[tool.poetry]
name = "digimon"
version = "0.1.0"
description = ""
authors = ["Nonsenj <oilchattong@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.112.0"}
sqlmodel = "^0.0.21"
pg8000 = "^1.31.2"
asyncpg = "^0.29.0"
pydantic-settings = "^2.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
